+++ celery/__init__.py
@@ -17,7 +17,7 @@ from . import local
 
 SERIES = 'dawn-chorus'
 
-__version__ = '5.2.1'
+__version__ = '5.2.3'
 __author__ = 'Ask Solem'
 __contact__ = 'auvipy@gmail.com'
 __homepage__ = 'http://celeryproject.org'
@@ -70,8 +70,7 @@ if STATICA_HACK:  # pragma: no cover
     from celery.app.base import Celery
     from celery.app.task import Task
     from celery.app.utils import bugreport
-    from celery.canvas import (chain, chord, chunks, group,  # noqa
-                               maybe_signature, signature, subtask, xmap,
+    from celery.canvas import (chain, chord, chunks, group, maybe_signature, signature, subtask, xmap,  # noqa
                                xstarmap)
     from celery.utils import uuid