+++ celery/exceptions.py
@@ -53,8 +53,7 @@ Error Hierarchy
 
 import numbers
 
-from billiard.exceptions import (SoftTimeLimitExceeded, Terminated,
-                                 TimeLimitExceeded, WorkerLostError)
+from billiard.exceptions import SoftTimeLimitExceeded, Terminated, TimeLimitExceeded, WorkerLostError
 from click import ClickException
 from kombu.exceptions import OperationalError