+++ t/unit/apps/test_multi.py
@@ -7,8 +7,7 @@ from unittest.mock import Mock, call, patch
 import pytest
 
 import t.skip
-from celery.apps.multi import (Cluster, MultiParser, NamespacedOptionParser,
-                               Node, format_opt)
+from celery.apps.multi import Cluster, MultiParser, NamespacedOptionParser, Node, format_opt
 
 
 class test_functions: