+++ t/unit/concurrency/test_prefork.py
@@ -66,9 +66,7 @@ class test_process_initializer:
     def test_process_initializer(self, _signals, set_mp_process_title, restore_logging):
         from celery import signals
         from celery._state import _tls
-        from celery.concurrency.prefork import (WORKER_SIGIGNORE,
-                                                WORKER_SIGRESET,
-                                                process_initializer)
+        from celery.concurrency.prefork import WORKER_SIGIGNORE, WORKER_SIGRESET, process_initializer
         on_worker_process_init = Mock()
         signals.worker_process_init.connect(on_worker_process_init)
 
@@ -343,6 +341,17 @@ class test_AsynPool:
 
         # Then: all items were removed from the managed data source
         assert fd_iter == {}, "Expected all items removed from managed dict"
+
+    def test_register_with_event_loop__no_on_tick_dupes(self):
+        """Ensure AsynPool's register_with_event_loop only registers
+        on_poll_start in the event loop the first time it's called. This
+        prevents a leak when the Consumer is restarted.
+        """
+        pool = asynpool.AsynPool(threads=False)
+        hub = Mock(name='hub')
+        pool.register_with_event_loop(hub)
+        pool.register_with_event_loop(hub)
+        hub.on_tick.add.assert_called_once()
 
 
 @t.skip.if_win32