+++ t/unit/utils/test_platforms.py
@@ -11,15 +11,11 @@ import pytest
 import t.skip
 from celery import _find_option_with_arg, platforms
 from celery.exceptions import SecurityError, SecurityWarning
-from celery.platforms import (ASSUMING_ROOT, ROOT_DISALLOWED,
-                              ROOT_DISCOURAGED, DaemonContext, LockFailed,
-                              Pidfile, _setgroups_hack, check_privileges,
-                              close_open_fds, create_pidlock, detached,
-                              fd_by_path, get_fdmax, ignore_errno, initgroups,
-                              isatty, maybe_drop_privileges, parse_gid,
-                              parse_uid, set_mp_process_title, set_pdeathsig,
-                              set_process_title, setgid, setgroups, setuid,
-                              signals)
+from celery.platforms import (ASSUMING_ROOT, ROOT_DISALLOWED, ROOT_DISCOURAGED, DaemonContext, LockFailed, Pidfile,
+                              _setgroups_hack, check_privileges, close_open_fds, create_pidlock, detached,
+                              fd_by_path, get_fdmax, ignore_errno, initgroups, isatty, maybe_drop_privileges,
+                              parse_gid, parse_uid, set_mp_process_title, set_pdeathsig, set_process_title, setgid,
+                              setgroups, setuid, signals)
 from celery.utils.text import WhateverIO
 from t.unit import conftest